To get rid of Visual C++ warnings explicit type casts were added.
svn path=/trunk/; revision=9728

+++ drivers/svg.c
@@ -962,7 +962,7 @@ void write_unicode(FILE *svgFile, PLUNICODE ucs4_char)
 short desired_offset(short level, double ftHt)
 {
   if(level != 0){
-    return -level * ftHt * pow(0.6, abs(level));
+    return (short)(-level * ftHt * pow(0.6, abs(level)));
   }
   else{
     return 0;
